replace single with double quote and uint with uint256

+++ contracts/interfaces/IOracle.sol
@@ -2,6 +2,6 @@
 pragma solidity >=0.8.0;
 
 interface IOracle {
-    function consult(address tokenIn, uint amountIn) external view returns (uint amountOut);
+    function consult(address tokenIn, uint256 amountIn) external view returns (uint256 amountOut);
     function update() external;
 }
